# docs/api/unified-api-spec.yaml
openapi: 3.0.3
info:
  title: SPARC Platform Unified API
  description: |
    Comprehensive API specification for the SPARC (Security Platform with Advanced Real-time Control) system.
    This unified specification covers all 24 microservices providing access control, video management, 
    environmental monitoring, visitor management, mobile credentials, and compliance features.
    
    ## Key Features
    - Multi-tenant architecture with tenant isolation
    - Real-time event processing and notifications
    - Offline resilience with 72-hour operation capability
    - Scalable to 10,000 doors and 1,000 concurrent video streams
    - SOX, HIPAA, and PCI-DSS compliance
    - Mobile credential management with mesh networking
    - Advanced analytics and reporting
    
    ## Authentication
    All endpoints require JWT authentication with tenant-specific scoping.
    Include the `Authorization: Bearer <token>` header and `X-Tenant-ID` header in all requests.
    
    ## Multi-Tenant Usage
    All API endpoints are tenant-scoped. Include the `X-Tenant-ID` header to specify the tenant context.
    Data isolation is enforced at the database and application levels.
  version: 1.0.0
  contact:
    name: SPARC Platform API Support
    email: api-support@sparc-platform.com
  license:
    name: Proprietary
    url: https://sparc-platform.com/license
servers:
  - url: https://api.sparc-platform.com/v1
    description: Production API Server
...
    description: Staging API Server
...
security:
  - BearerAuth: []
  - TenantHeader: []
paths:
...
  /video/cameras/{cameraId}/stream:
    get:
      tags:
        - Video Management
      summary: Get camera live stream URL
      parameters:
        - $ref: '#/components/parameters/CameraIdParam'
        - name: quality
          in: query
          schema:
            type: string
            enum: [low, medium, high, ultra]
            default: medium
      responses:
        '200':
          description: Stream URL retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  streamUrl:
                    type: string
                    format: uri
                  protocol:
                    type: string
                    enum: [rtsp, webrtc, hls]
                  expiresAt:
                    type: string
                    format: date-time
  /video/recordings:
    get:
      tags:
        - Video Management
      summary: List video recordings
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: cameraId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: eventType
          in: query
          schema:
            type: string
            enum: [motion, access_event, alarm, manual]
      responses:
        '200':
          description: Recordings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingListResponse'
  /video/recordings/{recordingId}:
    get:
      tags:
        - Video Management
      summary: Get recording details and playback URL
      parameters:
        - name: recordingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recording details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recording'

  # Environmental Monitoring Service
...
  /security/compliance/reports:
    get:
      tags:
        - Security & Compliance Service
      summary: List compliance reports
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: framework
          in: query
          schema:
            type: string
            enum: [sox, hipaa, pci_dss, gdpr, ccpa, iso27001]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending, completed, failed]
      responses:
        '200':
          description: Compliance reports retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceReportListResponse'

    post:
      tags:
        - Security & Compliance Service
      summary: Generate compliance report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateComplianceReportRequest'
      responses:
        '202':
          description: Compliance report generation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  reportId:
                    type: string
                  status:
                    type: string
                    enum: [pending, generating, completed, failed]

  # Maintenance Service
...
  /events/stream:
    get:
      tags:
        - Real-time Events
      summary: WebSocket endpoint for real-time events
      description: |
        WebSocket endpoint for receiving real-time events. 
        Upgrade to WebSocket protocol to receive live events.
        
        Event types include:
        - access_event: Door access attempts
        - video_alert: Video analytics alerts
        - environmental_alert: Environmental sensor alarms
        - system_alert: System health notifications
        - visitor_event: Visitor check-in/out events
        - maintenance_alert: Maintenance notifications
        - security_incident: Security incidents
        - backup_status: Backup job status updates
        - integration_event: Integration sync events
        
        Authentication required via query parameter: ?token=<jwt_token>
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: JWT authentication token
        - name: eventTypes
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter events by type
      responses:
        '101':
          description: Switching Protocols to WebSocket
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Health Check
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Tenant identifier for multi-tenant isolation
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
...
    CameraIdParam:
      name: cameraId
      in: path
      required: true
      schema:
        type: string
      description: Unique camera identifier
  schemas:
...
    Recording:
      type: object
      properties:
        id:
          type: string
        cameraId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in seconds
        fileSize:
          type: integer
          description: File size in bytes
        eventType:
          type: string
          enum: [motion, access_event, alarm, manual, scheduled]
        playbackUrl:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        metadata:
          type: object
          properties:
            resolution:
              type: string
            frameRate:
              type: integer
            codec:
              type: string
    RecordingListResponse:
      type: object
      properties:
        recordings:
          type: array
          items:
            $ref: '#/components/schemas/Recording'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    # Environmental Monitoring Schemas
...
    ComplianceReport:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        framework:
          type: string
          enum: [sox, hipaa, pci_dss, gdpr, ccpa, iso27001]
        status:
          type: string
          enum: [draft, pending, completed, failed]
        reportPeriod:
          type: object
          properties:
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
        findings:
          type: array
          items:
            type: object
            properties:
              control:
                type: string
              status:
                type: string
                enum: [compliant, non_compliant, not_applicable]
              evidence:
                type: array
                items:
                  type: string
              recommendations:
                type: array
                items:
                  type: string
        generatedAt:
          type: string
          format: date-time
        generatedBy:
          type: string
    GenerateComplianceReportRequest:
      type: object
      required:
        - framework
        - reportPeriod
      properties:
        framework:
          type: string
          enum: [sox, hipaa, pci_dss, gdpr, ccpa, iso27001]
        reportPeriod:
          type: object
          properties:
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
        includeEvidence:
          type: boolean
          default: true
        format:
          type: string
          enum: [pdf, excel, json]
          default: pdf
    ComplianceReportListResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/ComplianceReport'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    # Maintenance Service Schemas
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Authentication required
            timestamp: "2024-01-15T10:30:00Z"
            requestId: req_123456789
...